+++ decentraland/screens/ListLandForSaleScreen.js
@@ -8,9 +8,10 @@ import {
 import PropTypes from "prop-types";
 import LandForSaleList from "@presentational/LandForSaleList";
 import LandForSaleListItem from "@presentational/LandForSaleListItem";
-import { showError, showInfo } from "@helpers";
-import { getAllAssetsForSale } from "@helpers/decentraland";
+import { showError, showInfo, getContracts, addressesAreEqual } from "@helpers";
+import { generateAssetFromId } from "@helpers/decentraland";
 import { Root } from "native-base";
+import DecentralandUtils from "tasit-sdk/dist/helpers/DecentralandUtils";
 
 export class ListLandForSaleScreen extends React.Component {
   componentDidMount = async () => {
@@ -20,7 +21,7 @@ export class ListLandForSaleScreen extends React.Component {
     } = this.props;
     try {
       showInfo("Loading land for sale...");
-      const assetsForSale = await getAllAssetsForSale();
+      const assetsForSale = await this._getAllAssetsForSale();
       for (let promise of assetsForSale) {
         const assetForSale = await promise;
         appendLandForSaleToList(assetForSale);
@@ -29,6 +30,80 @@ export class ListLandForSaleScreen extends React.Component {
     } catch (err) {
       showError(err);
     }
+  };
+
+  // Note: Returns a list of Promises
+  _getAllAssetsForSale = async () => {
+    const decentralandUtils = new DecentralandUtils();
+    const { getAllAssetsForSale: getAllOpenSellOrders } = decentralandUtils;
+
+    const openSellOrders = await getAllOpenSellOrders();
+
+    const contracts = getContracts();
+    const { estateContract } = contracts;
+
+    const estatesForSale = [];
+    for (let order of openSellOrders) {
+      const { nftAddress } = order;
+      const isEstate = addressesAreEqual(
+        nftAddress,
+        estateContract.getAddress()
+      );
+      if (isEstate) estatesForSale.push(order);
+    }
+
+    const assetsForSale = [];
+    // Note: Getting only the first 10 assets for now
+    // See more: https://github.com/tasitlabs/tasit/issues/155
+    const listSize = 10;
+    for (let order of estatesForSale.slice(0, listSize)) {
+      let assetForSalePromise = this._toAssetForSale(order);
+      assetsForSale.push(assetForSalePromise);
+    }
+
+    return assetsForSale;
+  };
+
+  _toAssetForSale = async sellOrder => {
+    const contracts = getContracts();
+    const { estateContract, landContract } = contracts;
+    const {
+      id,
+      nftAddress,
+      assetId,
+      seller,
+      priceInWei,
+      expiresAt,
+    } = sellOrder;
+
+    // Note: Conversion to USD will be implemented on v0.2.0
+    const manaPerUsd = 30;
+    // Get mana price using string to avoid imprecise rounding (i.e.: 57999.99999999999)
+    // TODO: Use TasitSDK Utils to dealing with BigNumbers (will be implemented on v0.2.0)
+    const priceManaInWei = `${priceInWei}`;
+    const strPriceManaLength = priceManaInWei.length - 18;
+    const strRoundedPriceMana = priceManaInWei.substring(0, strPriceManaLength);
+    const priceMana = strRoundedPriceMana;
+    const priceUSD = Number(priceMana / manaPerUsd).toFixed(2);
+
+    const asset = await generateAssetFromId(
+      estateContract,
+      landContract,
+      assetId,
+      nftAddress
+    );
+
+    const assetForSale = {
+      id,
+      priceManaInWei,
+      priceMana,
+      priceUSD,
+      seller,
+      expiresAt,
+      asset,
+    };
+
+    return assetForSale;
   };
 
   _renderItem = ({ item: landForSale }) => {
