+++ decentraland/screens/MyAssetsScreen.js
@@ -4,9 +4,10 @@ import { setMyAssetsList } from "../redux/actions";
 import PropTypes from "prop-types";
 import MyAssetsList from "@presentational/MyAssetsList";
 import MyAssetsListItem from "@presentational/MyAssetsListItem";
-import { listsAreEqual } from "@helpers";
-import { getAssetsOf } from "@helpers/decentraland";
+import { listsAreEqual, getContracts } from "@helpers";
+import { generateAssetFromId } from "@helpers/decentraland";
 import { storeMyAssets } from "@helpers/storage";
+import DecentralandUtils from "tasit-sdk/dist/helpers/DecentralandUtils";
 
 export class MyAssetsScreen extends React.Component {
   componentDidMount = async () => {
@@ -26,9 +27,36 @@ export class MyAssetsScreen extends React.Component {
   };
 
   _getAssetsFromBlockchain = async address => {
-    const listOfPromises = await getAssetsOf(address);
+    const listOfPromises = await this._getAssetsOf(address);
     const fromBlockchain = await Promise.all([...listOfPromises]);
     return fromBlockchain;
+  };
+
+  // Note: Returns a list of Promises
+  _getAssetsOf = async address => {
+    const decentralandUtils = new DecentralandUtils();
+    const { getAssetsOf: getLandOf } = decentralandUtils;
+
+    const contracts = getContracts();
+    const { estateContract, landContract } = contracts;
+
+    const listOfLand = await getLandOf(address);
+    const assets = [];
+
+    for (let land of listOfLand) {
+      const { id: assetId, nftAddress } = land;
+
+      let asset = generateAssetFromId(
+        estateContract,
+        landContract,
+        assetId,
+        nftAddress
+      );
+
+      assets.push(asset);
+    }
+
+    return assets;
   };
 
   _renderItem = ({ item: asset }) => {
